Move ConsensusParams back to original file
